fix(api-markdown-documenter): Always display type param info when it is present (#19466)
The code as written would only display type parameter information for
`class` and `interface` items. This was an oversight when making the
initial implementation. The code now correctly displays type parameter
docs in the `Signature` section if the API item has any.

Also updates type parameter tables to display type constraint and
default excerpts, instead of an "optional" notice.

See the updated test output for an example of the changes.

### Breaking Change

These changes also update the signature of `createTypeParametersSection`
to be a bit more user friendly, which is at least semantically breaking.
It will now _always_ generate the section if asked. If the consumer
wants to omit the section (for example when the list of type parameters
is empty), they can make the call conditional on their end. The benefit
is that consumers no longer have to make checks for undefined results.

### Misc

Also resolves an issue with the build pipeline for this package. Since
this package uses a `file:` dev dependency on
`@fluidframework/eslint-config-fluid`, we need to run the build pipeline
when changes are made in that directory.

AB#7033