GC: Auto-Recovery triggers fullGC to self-repair cases of corrupted GC Data (#19392)
Follow-on after #19115

### Background

With that first Auto-Recovery PR, history is doomed to repeat itself -
meaning, even though clearing the unref state saves the object from
deletion, it'll likely come back as Tombstoned again a few weeks later.

If the bug is on the app side (somehow they're requesting an object
that's not properly referenced), then the ball is in their court and
there's nothing we can do on FF side.

But if the object _actually is referenced_, but GC miscalculated the
reference graph (e.g. GC Data was corrupted by a summarizer corner
case), then we have a chance to do something about it.

### The Fix

When processing the "TombstoneLoaded" op for auto-recovery, in addition
to clearing the unreferenced state tracking, also ensure that on the
next GC run, `fullGC` mode will be used. This will cause every node to
re-inspect itself to find the set of outbound routes it contains, hence
correcting any latent errors in the graph passed to the core GC
algorithm.

### Test Plan

* Unit tests to cover the new flag
* End-to-end test similar to the existing auto-recovery test. But
instead of accessing the Tombstoned object via faking a request, we
manually corrupt the GC Data so that even `handle.get` results in a
Tombstone error 😱