build(example-utils): Add explicit return type to getter (#19480)
api-extractor/typescript do not guarantee that the emitted order of an
_inferred_ string-literal union type is stable. This can lead to
"flip-flopping" where api-extractor sometimes emits `"foo" | "bar"` and
sometimes `"bar" | "foo"`.

Adding explicit types instead of relying on inference is the best
workaround. In this case there was already an exported type defined so I
used that instead of an explicit literal union.