TableBench: use 'object of arrays' for baseline (#19386)
One of the things TableBench measures is the ratio of "actual data" to
the resulting summary size.

When calculating the size of the "actual data", we used the 'array of
objects' form returned by the major databases. This, however, repeats
the column keys per row in the JSON encoding.

For a more competative interpretation of "actual data", we now transpose
the data into 'object of arrays' form, where each column key appears
exactly once followed by a dense array of values.