(cherry-pick 146fcaa90e75ba1ce17ea2e3cecaf85ebaf45cd9) (cherry-pick 15bce6515457548f99cfe8f439be1f512e96763f)
GC: Properly support running GC only for certain node types (#19524)

Fixes AB#7141

When implementing two-stage sweep (including follow-up bug fix #18734),
we missed the case where config flags are set to disable GC Sweep for
either blobs or datastores. In those cases, the excluded nodes should be
tombstoned, but as it is they will not be.

The fix involves:

* Removing the option to disable blob GC. It's the simpler case, and I
don't anticipate needing this ability. It reduces the support matrix for
incoming configs, which outweighs the loss of functionality.
* Consolidating the logic for deciding what nodes to run sweep for into
the GC classes, rather than out in other parts of the runtime.
* Adding a new `ONLY_BLOBS` case to the `shouldRunSweep` internal config