Finalize ID Range on Attach and Allow ContainerRuntime Deletion (#19560)
This PR adds some functionality needed for #19400. We do not currently
finalize the id creation ranges on attach, which will prevent the ID
compressor from being able to function when rehydrating from a summary
is allowed. We also don't have a way to completely remove a container
runtime from a factory, which is needed to ensure that any references to
the old client are gone once we rehydrate from a summary.

---------

Co-authored-by: Tony Murphy <anthony.murphy@microsoft.com>