fix(end-to-end-tests): Add AgentScheduler to DataRuntime APIs to support back compat tests (#19573)
## Description

This PR modifies the following in order to enable cross version compat
testing for AgentScheduler:

1. AgentScheduler is added to the DataRuntime APIs. This gives the e2e
test suite access to the legacy `AgentSchedulerFactory`
2. `TestObjectProviderWithVersionedLoad` no longer alternates between
creating/loading versions. This simplifies the flow and makes it easier
for test writers to use the correct version when providing their own
runtime factories.
3. The AgentScheudler test suite provides the proper runtime factory
based on if we are creating/loading containers in order to match
`TestObjectProviderWithVersionedLoad` (this only applies for cross
version compat tests).