Align apply stashed ops with re-submit (#19518)
## Background
This change is in preparation for allowing consumers of fluid to call
serialize on a container while a container is in an attaching state,
which will include a failed attach where the failure didn't cause the
container to close. This will aid in the case of offline, network, or
server outage by allowing the application to capture the un-attached
state of the container, preserve it locally, and later create a new file
from it.

AB#5502

## Overview
Similar to capturing pending state we have two distinct means for
applying pending state. For detached containers we simply reload from
the captured summary/snapshot and continue using the container in the
detached state. For attached containers we similarly reload from the
same snapshot that was originally loaded and the container remain
attached. However, we also have the pendingRuntimeState, which includes
remote and local operations which need to be replayed on reload. After
reloading the container, it will catch up with new remote operations on
the server, rebase the local operations applied from the pending state,
and then send those operations to the server.

As discussed, above about capturing pending state, the attaching state
will be the same as the detached state but also include the
pendingRuntimeState. The pendingRuntimeState can be processed much like
it is processed in that attached state. Specifically, the
pendingRuntimeState will get passed to the ContainerRuntime which will
enumerate the pending local state contained in it, and pass the
contained ops to the a ContainerRuntime.applyStashedOps. applyStashedOps
is similar to all our op processing paths, like those for processing
remote ops, rollback, and resubmit. Each of these start at the
ContainerRuntime, then move through the DataStoreRuntime, and finally go
to the IChannel, which are DDS. Each layer may process the op if that
layer owns it or pass it along to the layers below it.

Today these processing paths do not reach all the way to DDS in the
detached case, as these paths are not generally used when detached. This
will have to change to support calling applyStashedOps while detached.
Specifically, the following changes will be needed:
•	Add support for applyStashedOps to RehydratedLocalChannelContext
• Change SharedObject to attachDeltaHandler, which is how
applyStashedOps path connects to SharedObjects, when not attached.

After the above changes, it will also be necessary for SharedObjects to
support applyStashedOps while in a detached state. There are multiple
ways to support this:

1. Change apply stashed ops such that dds return “undefined”, or
“Symbol.For(..)” when detached. Handle that in the pending state manager
and don’t consider the op is submitted. The key here is dds need to
understand that these applied ops will never be acknowledged, so they
should not generate local op meta data.
2. Change apply stashed ops such that dds need to call submit, like in
resubmit, and not automatically considering stashed ops submitted like
we do today. In general, this allows dds to reuse their modification
apis, rather than needing custom implementations for apply stashed ops.

Option 2 is the preferred path, as it both support the new path more
easily, but also simplifies working with local state in that attached
case. This is implemented for the existing applyStatshedOp path.

The change isolates the handling of submitting ops during
applyStatshedOp in the channel delta connection use for shared objects.
Additionally, i've moved a number of DDS to the new mode to illustrate
the gains with approach (lots of deleted code). The rest of the DDS will
be modified in follow up changes.

Test coverage for this change is provided by existing stashed ops
coverage.


[AB#7143](https://dev.azure.com/fluidframework/internal/_workitems/edit/7143)

---------

Co-authored-by: Tony Murphy <anthonm@microsoft.com>