tree: Fix bug with cross-field move invalidation (#19571)
## Description

Previously ModularChangeFamily registered cross-field dependencies for a
field after running compose, rebase, or invert on that field. This meant
that a field could not be invalidated by cross-field information
provided by a descendant field. This change fixes this issue by
registering dependencies immediately instead of as a batch after
finishing processing the field.

---------

Co-authored-by: yann-achard-MS <97201204+yann-achard-MS@users.noreply.github.com>