Redo summarizer correctness validation (#19488)
[AB#3400](https://dev.azure.com/fluidframework/internal/_workitems/edit/3400)

Updated from previous PR:
https://github.com/microsoft/FluidFramework/pull/18754 as it was getting
stale.

## Description
After ensuring that we've refreshed our latest summary state:
https://github.com/microsoft/FluidFramework/pull/18968, this PR adds
validation that before we start a summary, that all the summary sequence
numbers are as expected from the `RunningSummarizer` down to the
`SummarizerNode` system itself. This is to prevent any sort of data
corruption/data loss from accidentally occurring. This check used to run
in the two-commit summary system, which would prevent us from submitting
a bad summary. Now that we've moved to the one-commit summary system,
this should keep us safe from potentially negatively impacting
documents.

## What's in the PR
This PR adds validation that the summary system from the
RunningSummarizer down to the SummarizerNode system that the starting
sequence numbers are correct.

The code is behind a feature flag. There is telemetry that's added so
that we can see before we even turn on the feature what the impact to
the entire system will look like.

---------

Co-authored-by: Mark Fields <markfields@users.noreply.github.com>