build(docs): Don't generate API docs for `@alpha` members (#19605)
We reserve `@alpha` for use by existing partners, they are not intended
for public use. This PR updates the API docs generation to not emit
documentation entries for `@alpha` APIs.

This was previously blocked on ensuring our tagging was in the correct
shape across the repo, and ensuring that our manually written docs
didn't reference any such APIs. That work has now been completed, and
this change can be made safely.

AB#6469