[Data Virtualization]Upload groupId in summary plus other groupId conversions (#19587)
[AB#7061](https://dev.azure.com/fluidframework/internal/_workitems/edit/7061)

## Data Virtualization
Data virtualization is an effort where we introduce the concept of
`groupId`. Group id is on each data store and identifies whether or not
we want to make a request to the server to get that datastore. The main
datastores should not have groupId on the snapshot.

## Description
From a server and service standpoint, we just want to make sure
`groupId` is in the summaries. There's some testing that needs to be
done to figure out the exact plumbing to make sure `groupId` is
persisted in snapshots in both the `local` server and `tinylicious`
server. I had reviewed `unreferenced`, and at least in local server, it
did not round trip in the test. That part is commented out.

## Testing
- round trip of groupId through the `attachSummary` flow works
- submitted summaries contain `groupId`

## Future work
- add `groupId`to server round trip
- add testing to see that loading from a new summary/snapshot round
trips `groupId`