Re-enable migration shim tests (#19623)
[AB#6416](https://dev.azure.com/fluidframework/internal/_workitems/edit/6416)

Re-enables these data migration shim tests.

- `migrationShim.spec.ts`: added a `provider.ensureSynchronized` that
would send the handle and attach ops for the second container to load
from.
- `sharedTreeShim.spec.ts`: added a `provider.ensureSynchronized` that
would send the handle and attach ops for the second container to load
from. It also added a `provider.ensureSynchronized` to ensure that
`tree.schematize` would work as expected.
- `stampedV2Ops.spec.ts`: similar to `sharedTreeShim.spec.ts` above
- `storingHandlesDetached.spec.ts`: added a
`provider.ensureSynchronized` that would send the handle and attach ops
for the second container to load from. Also added a
`provider.updateDocumentId` call so the second load container would
successfully load the container.