Remove provider.ensureSynchronized to remove flakiness of test (#19626)
[AB#6706](https://dev.azure.com/fluidframework/internal/_workitems/edit/6706)

## Description
Reduce flakiness of test.

What was happening was we were expecting to process a summary ack that
would close the container. The test would hang because sometimes we
would process both acks before the summarizer would get to it and the
container would close, but we would run the summarizer again, which
wouldn't close the container. If we run the summarizer before we process
the ack, which was what the test originally intended, then we will hit
our failure scenario.