refactor(api-markdown-documenter): Adapt `IndentedWriter` from `@microsoft/api-documenter`, rather than consuming external package internals (#19472)
This library was originally built on top of `api-documenter`, but was
slowly migrated off of it. There are still a couple of lingering
references to `api-documenter` internals, which is bad. One of these is
their `IndentedWriter`, which is a stateful string builder that handles
pretty indented formatting.

Since both packages are under MIT license, I went ahead and adapted
their implementation to our needs and moved it into our package. A
couple of methods were removed and/or made private due to our more
limited needs.

### Breaking Change

This PR changes how a `DocumentWriter` is instantiated. Rather than
exposing a free function (`createDocumentWriter`), we now expose a
`create` method in a namespace with the same name as the new interface
(`DocumentWriter.create`).