Decouple DDS isAttached from services (#19590)
## Background
This change is in preparation for allowing consumers of fluid to call
serialize on a container while a container is in an attaching state,
which will include a failed attach where the failure didn't cause the
container to close. This will aid in the case of offline, network, or
server outage by allowing the application to capture the un-attached
state of the container, preserve it locally, and later create a new file
from it.


[AB#5502](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/5502)

## Overview
As part of serializing while attaching we will also need to support
calling applyStashedOps on a detached container. In order to do this is
necessary for SharedObject to have an attached delta handler. Before
this change, the existence of the services and its handlers was also
used to determine if the dds itself was attached. This was conceptually
and functionally incorrect, as the attach state of a dds is determined
by two facts:

1. Is it bound to a context which means its handle is stored in an
already bound dds
2. Is the runtime attached

if both of these are true, then the dds should be considered attached.

Beyond attach, there is also the connected state, and the didAttach
method. The connect state should always be false while detached, and
didAttach should only be called one when the dds transitions.

This change tightens up the handling of the bind, attached, and connects
state of shared object while decoupling it from the services, to allow
the services to be set sooner. Lastly, this change adds a ton of
previously missing unit test coverage to ensure all the states are
managed properly.



[AB#7175](https://dev.azure.com/fluidframework/internal/_workitems/edit/7175)

---------

Co-authored-by: Tony Murphy <anthonm@microsoft.com>
Co-authored-by: Alex Villarreal <716334+alexvy86@users.noreply.github.com>