Enable groupId in local server base changes (#19662)
[AB#7196](https://dev.azure.com/fluidframework/internal/_workitems/edit/7196)

## Description
Enable local-driver and local-server to round trip `loadableGroupId`.
The gist is that we create a blob and store `groupId` in the upload
manager. (maybe we need to do it also on attach). On load we take that
blob and convert it to `groupId` on the snapshot. Not sure what the best
way to do this as this is more a POC currently than a final
implementation.

## GroupId
The `loadableGroupId` effort is intended to make it so data stores can
be virtualized. This is part of the data virtualization effort.

## Alternatives
Looked into enabling `WholeSummaryUpload`, which is not supported by
`TestHistorian` currently. Thus, we would need to build it out.