fix(id-compressor): Fix corrupting resubmit bug (#19723)
## Description

This change fixes a bug where resubmission of an ID allocation op was
incorrectly routed to the main batch manager. A DDS that generates IDs
during resubmit will result in allocation ops in both the allocation
batch manager and the main batch manager, causing ranges to be finalized
out of order and permanently corrupting the document. This change adds
asserts to prevent incorrect op types in specific batch managers.