[Data Virtualization] Add capability to load omitted snaphot in RemoteFluidDataStore (#19720)
## Description

This adds capability to load individual datastores with a loadingGroupId
if the snapshot content is omitted.
1.) Add function in container Runtime to fetch the snapshot based on
loadingGroupid and return the part of that snapshot for the path
requested. The group snapshot is cached so that it can used by used by
other requests. We use PromiseCache so that we can get the same promise
if it is not yet completed.
2.) The above API returns the snapshot when the container has caught up
to the snapshot seq number of the snapshot of a loading group. If it is
summarizer, then it throws an error as that means summarizer is not up
to date.
3.) In datastoreContext, when the snapshot is returned, we only process
the ops which are after the snapshot seq number.
4.) Also added various unit tests to cover different scenarios.

This is part of data virtualization effort which is documented here:
[Design
Doc](https://microsoft.sharepoint.com/:w:/t/FluidFramework/ESJQvukxffJNiTZzbNCWky4B891mpOpWD7BHhbOznXsMZg?e=gAUCrx)

---------

Co-authored-by: Jatin Garg <jatingarg@Jatins-MacBook-Pro-2.local>