appendLocatorParam function in odspURLResolverForShareLink (#19648)
- New utility function in `odspUrlResolverForShareLink` class of
odsp-driver which can add locator parameter (`nav` param) to the base
url provided as an input to the function.
- One of the immediate use-cases for this utility is that when the host
application has a sharing link (say generated from using the
[`SingleRT`](https://github.com/microsoft/FluidFramework/pull/11177)
feature), the host application can create a re-loadable link to the
fluid component by calling this utility. Up until now, the host apps had
to rely on the `getAbsoluteUrl` method, which was not ideal as it made a
network call to odsp to get the absolute url for the resource.
- Added tests as well to test the complete resolvedUrl information is
preserved in the `nav` param, and to test that no other params are
disturbed in the original url provided.