[Data Virtualization] Enable local server to virtualize parts of the snapshot (#19767)
[AB#7197](https://dev.azure.com/fluidframework/internal/_workitems/edit/7197)

## Description

Enables testing of local server to virtualize parts of the snapshot via
loading groupId. This proves that we can virtualize the snapshot and get
less data than what was originally there.

## Details

The changes were made in the local-driver to simulate the service
stripping the relevant data. We could try to do a full implementation,
but that would be more expensive than simply just doing it in one spot.

## Testing
- The snapshot includes the non groupId trees on the regular getSnapshot
call
- The snapshot excludes the groupId trees on the regular getSnapshot
call
- The snapshot includes the groupId trees on the groupId getSnapshot
call
- The snapshot excludes the non groupId trees on the groupId getSnapshot
call
- Can request remote virtualized groupId datastores/dataobjects.
- Remote contexts with groupId contain an empty base snapshot
- Network calls return expected snapshots.