Encapsulate container state attributes (#19697)
There is a lot of code spread out in container.ts related to capturing
the pending state and managing IPendingContainerState interface and
related attributes. We can abstract all that manipulation into a
separate class to simplify the container code and make future features
like refreshing the base snapshot unit testable.