[local-server bump]: Update client packages to consume local server detached groupId flow (#19831)
[AB#7249](https://dev.azure.com/fluidframework/internal/_workitems/edit/7249)

Enables the test `"Can create loadingGroupId via detached flow"` in
`groupIdInSummary.spec.ts`

Consume the changes produced on the `TestDocumentStorage` from this PR
https://github.com/microsoft/FluidFramework/pull/19743.

Essentially it enables groupId to round trip via the `container.attach`
flow.

For the **Data Virtualization Testing Effort**, this enables us to test
detached containers with local changes that include groupId changes. It
turns out that the `createDocument` flow and the `uploadSummary` flow in
local-server go through slightly different flows that both would drop
groupId. This is the update to the `createDocument` flow such that
`groupId` is persisted on the server and also is round-tripped back to
the clients.