Support applyingStashedOps while detached (#19802)
## Background
This change is in preparation for allowing consumers of fluid to call
serialize on a container while a container is in an attaching state,
which will include a failed attach where the failure didn't cause the
container to close. This will aid in the case of offline, network, or
server outage by allowing the application to capture the un-attached
state of the container, preserve it locally, and later create a new file
from it.


[AB#5502](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/5502)

## Overview

The majority of these changes are test changes which simulate having DDS
enter the attaching state, capturing their state, and then rehydrating
to a new instance of the DDS.

This was made possible with #19518. As that change made DDS apply
stashed ops via the DDSes existing functions to apply local changes, and
since those existing functions already supported the detached state,
there were no additional per DDS modifications necessary to apply
stashed ops on a rehydrated detached container.

Beyond the test changes this change makes the attachment state visible
to the apply stashed op pipeline, so it can know to not expect new ops
to be generated when detached, and not populate the pending state
manager.

The next PR
([AB#7077](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/7077))
of this work will enable capturing the pending state while detached.
That change will bring end to end test coverage that can't be done
currently. However, there is substantial coverage for the internals in
this PR via the integration with the stress harness.


[AB#7078](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/7078)

---------

Co-authored-by: Tony Murphy <anthonm@microsoft.com>
Co-authored-by: alex-pardes <105307164+alex-pardes@users.noreply.github.com>
Co-authored-by: Connor Skees <39542938+connorskees@users.noreply.github.com>