Enable calling serialize while container is in the attaching state (#19893)
## Background
This change is the final change for allowing consumers of fluid to call
serialize on a container while a container is in an attaching state,
which will include a failed attach where the failure didn't cause the
container to close. This will aid in the case of offline, network, or
server outage by allowing the application to capture the un-attached
state of the container, preserve it locally, and later create a new file
from it.


[AB#5502](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/5502)

## Overview
This change enables calling container.serialize while a container is in
the attaching state. It is the culmination of many smaller changes that
have gone in to enable this change.
- #18829
- #19400
 - #19246
 - #19517
 - #19518
 - #19590
 - #19634
 - #19802

---------

Co-authored-by: Tony Murphy <anthonm@microsoft.com>
Co-authored-by: Daniel Madrid <105010181+dannimad@users.noreply.github.com>
Co-authored-by: jzaffiro <110866475+jzaffiro@users.noreply.github.com>
Co-authored-by: Matt Rakow <ChumpChief@users.noreply.github.com>