fix(merge-tree sequence): slide references ad hoc inside localReferencePositionToPosition (#19780)
[AB#7109](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/7109)

The crux of the issue is that we don't slide references until after ack
of a remove, so when you call localReferencePositionToPosition it
returns the position of the removed segment. This is fine when we slide
to the right (forward) because we end up getting the correct position
anyway, except in the case when the end of the string is entirely
deleted. In this case and in the case of a (left) forward slide, the
returned position is greater by exactly the number of segments with
unacked removes that are adjacent to the front of the segment.