fix(matrix): Properly detect row/cell deletion when resubmitting a set op (#19862)
## Description

This adjusts the matrix resubmit logic for its `setCell` ops. Prior
changes didn't correctly rebase over row/column changes in a sensible
way.

The general problem here is that when resubmitting a setCell operation,
matrix must correctly resolve the cell's location using its row/col
permutation vectors. There was one simple thing wrong with this logic:
if the row or column for that cell has since been removed, there's
nothing to resubmit.

There's also another trickier bug: if the setCell operation was for an
unacked row or column, it's possible that while resubmitting the
insertion of that row/col, we discovered that it was inserted into a
removed range. In this case, for consistency across other clients,
merge-tree normalizes the location of the inserted segment to the start
of the removed range (that's where the insertion will happen on
non-local clients). This means that positional reasoning involving that
segment over the window between the original setCell op and the current
sequence number won't always resolve correctly.

This PR takes the approach of adding local reference positions on the
row/col segments of the setCell op for use during reconnect.

---------

Co-authored-by: Abram Sanderson <absander@microsoft.com>