Use DDS interfaces instead of concrete types in many tests and examples (#19936)
## Description

Concreate DDS types leak a lot of implementation details (like that they
extends SharedObject, SharedObjectCore and
EventEmitterWithErrorHandling. These are all details these use cases
don't need to depend on, and thus shouldn't depend on. Migrating to the
interfaces removes this unneeded dependency.

This is work toward reducing the scope in which the concrete DDS types
are exported, but also good standalone cleanup.