Increase maxCacheExpiryTimeoutMs to 5 days, and validate in FluidCache (#19997)
This change increases the default cache expiry timeout to the maximum
and renames it as such. It also moves the value to odsp-definitions so
that it can be reused by FluidCache. FluidCache uses the value to
validate that the user provided expiry time is below the maximum or logs
an error to inform the user.

---------

Co-authored-by: Tony Murphy <anthonm@microsoft.com>
Co-authored-by: Mark Fields <markfields@users.noreply.github.com>