TestHistorian.getBlob() to better handle sha = undefined calls. (#19987)
I run into a problem where due to a bug in my code I ended up hitting.
this function with sha === undefined.

Without this fix, we are getting some random blob content back, which is very hard to figure out, unless you debug this code. The issue is described in comment I'm adding - it's some weird expectations that I do not fully understand, so adding specific case that should be scoped to solve the problem at hand.