Keep pending checkpoint message for future summaries (#20029)
## Description

During a session, we might do multiple client/service summary, and
independently, multiple checkpoints. Checkpoint, will clear messages
storage in `pendingCheckpointMessages`, which is also used for writing
summary. Because of this cleanup, when write new summaries, summary
write often need to request the ops from alfred again, which is not
quite efficient.

So changed the behavior to keep `pendingCheckpointMessages` to max of:
1. protocolHead for efficient
2. max sn - max logtail length to avoid memory leak.

And used `lastCheckpointInsertedNumber` to preserve the backwards
compatibility when insert db (FRS not use DB)

---------

Co-authored-by: Xin Zhang <zhangxin@microsoft.com>