Add top-level transaction APIs to SharedTree (#20033)
## Description

This PR adds `Tree.runTransaction` to allow synchronous transactions to
be easily run from either a node or a tree view. This is a small
departure from the existing functions on the global `Tree` object:
1. It is the first non-read-only function
2. It requires types that are exposed in the shared-tree directory,
rather than solely in the simple-tree directory.

The shared-tree directory depends on simple-tree, not the other way
around, and therefore the `Tree` API needed to be lifted into
shared-tree from simple-tree to accommodate departure 2, above. I chose
to do this by creating a second `Tree` object in shared-tree that wraps
the `Tree` object in simple-tree and leaves the latter alone. This means
we don't have to expose many of the helper functions used by `Tree` in
simple-tree that are otherwise local to simple-tree and would leak up
into shared-tree. Accordingly, the `Tree` in simple-tree has been
renamed to just `treeNodeApi`, and the one in shared-tree is now the new
`Tree`. Tests that directly use `treeNodeApi` still alias it to `Tree`
because it is recognizable.

Also:
- I moved the `Tree.on` tests into the same file as the rest of the
simple-tree `Tree` tests (this gets rid of node.spec.ts).
- I lifted the `getView` test helper into shared-tree from simple-tree
since it's used by both treeApi.spec test files.

Support for constraints in transactions will come in a later PR.